Prepare cross references database from https://www.openbible.info/labs/cross-references/
The commit adds a sed script file (cr_to_db.sed) that maps bible book
abbreviations to their corresponding numbers and formats the data for
insertion into a SQLite database. Additionally, the Makefile is updated
to include rules for downloading the cross-references ZIP file,
extracting the data, and transforming it into a CSV file suitable for
importing into a SQLite database.