Add references to verses
In this commit, references to verses were added in the chapters.sql
file. A new condition was introduced to filter references with a rate
greater than 3. Additionally, in the same file, the logic related to
grouping verses was modified to differentiate between verses with
references and those without, affecting the structure and content of the
output.