Customize name of the book
The previous commit introduced modifications to the Makefile,
specifically altering the rate variable from 15 to 10, and adding a new
variable 'name' with the value 'pocket-nt'. These changes allow for the
customization of the book title displayed in the generated HTML output.
By updating the HTML generation process in the Makefile, users can now
easily specify the name of the book they are working on, offering a more
personalized touch to their projects.