Select references with CSS
This commit updates the CSS styling in the `style.css` file to include
specific selectors that will hide references based on their rates. By
adding CSS rules for references with rates ranging from 1 to 9, these
references will now be hidden when displayed on the web page. This
modification enhances the visual presentation of references for a better
user experience.