Fix print.css stylesheet for print version
The print.css stylesheet was updated to adjust the size of the root
element and the font size for the verse:before selector. These changes
were made to improve the styling and layout specifically for the print
version of the document. The root element size was increased from 2.1vh
to 2.3vh, and the font size for the verse:before selector was adjusted
from .6em to .5em for better readability and aesthetics in the printed
format.